added utility service layer for abstraction
