+++ backend/projects/SmartSocial/src/main/java/de/hsbi/smartsocial/Exceptions/RefreshException.java
@@ -1,0 +1,7 @@
+package de.hsbi.smartsocial.Exceptions;
+
+public class RefreshException extends RuntimeException {
+    public RefreshException(String message) {
+        super(message);
+    }
+}
